requirements.txt
fastapi
uvicorn
streamlit
python-multipart
requests
python-dotenv
pymongo
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
chromadb
sentence-transformers
faiss-cpu
edge-tts
aiohttp
paddleocr
paddlepaddle
PyMuPDF
pypdf
Werkzeug
pydantic
scikit-learn
beautifulsoup4
pillow
numpy
pysqlite3-binary

langchain-groq
